requirements.txt
requests
openai
anthropic
ollama
PyYAML 